feat(pickup in store on Checkout lint and config update)
