+++ feature-libs/pickup-in-store/components/container/pdp-pickup-options-container/pdp-pickup-options-container.component.ts
@@ -20,12 +20,12 @@ import {
   PickupOptionFacade,
   RequiredDeepPath,
 } from '@spartacus/pickup-in-store/root';
+import { getProperty } from '@spartacus/pickup-in-store/utils/utils';
 import {
   CurrentProductService,
   LaunchDialogService,
   LAUNCH_CALLER,
 } from '@spartacus/storefront';
-import { getProperty } from 'feature-libs/pickup-in-store/utils/utils';
 import { combineLatest, iif, Observable, of, Subscription } from 'rxjs';
 import { filter, map, startWith, switchMap, take, tap } from 'rxjs/operators';
