+++ feature-libs/pickup-in-store/core/utils/utils.ts
@@ -17,3 +17,16 @@ export function isInStock(stockInfo: Stock | undefined): boolean {
 export function storeHasStock({ stockInfo }: PointOfServiceStock): boolean {
   return isInStock(stockInfo);
 }
+
+export const getProperty = (
+  o: Object | undefined | null,
+  property: string
+): any | null => {
+  if (!o) {
+    return null;
+  }
+  if (o.hasOwnProperty(property)) {
+    return o[property];
+  }
+  return null;
+};
