+++ feature-libs/pickup-in-store/root/model/delivery-points.model.ts
@@ -8,9 +8,7 @@ import { OrderEntry } from '@spartacus/cart/base/root';
 import { PointOfService } from '@spartacus/core';
 
 export type DeliveryPointOfServiceItems = {
-    name: string;
-    storeDetails: PointOfService;
-    value: Array<OrderEntry>;
-}
-
-
+  name: string;
+  storeDetails: PointOfService;
+  value: Array<OrderEntry>;
+};
