+++ feature-libs/pickup-in-store/root/model/index.ts
@@ -7,4 +7,4 @@
 export * from './augmented-core.model';
 export * from './pickup-option.model';
 export * from './stock-location-search-params.model';
-export * from './delivery-points.model'
+export * from './delivery-points.model';
